src/examples/mlviz/requirements.txt
umap-learn
lightgbm
xgboost
pandas
scikit-learn
tqdm